requirements.txt
Django>=4.2
gunicorn>=21.0
djangorestframework>=3.16.0